Import zsh completition made by Holger Macht
Right now the zsh completition is packaged inside the zsh package on
OpenSUSE.
Several distributions either package it or the user has to import it
themselves.

It is better if it is stored here so any can pick it up in there zsh
package and put fixes in here.

The file was imported from the OpenSUSE zsh package plus a few spelling fixes.

Signed-off-by: Björn Bidar <bjorn.bidar@jolla.com>